requirements.txt
Flask
Flask-SQLAlchemy
postgres
psycopg2-pool
python-dotenv
SQLAlchemy
flask
psycopg2-binary
